# contracts/openapi/dx-os-v1.yaml
openapi: 3.1.0
info:
  title: DX-OS API
  version: 0.8.0
  description: Contract cho Go API, Procurement, Budget, Attachment và Reporting của DX-OS.
  contact:
    name: DX-OS Engineering
    url: https://docs.dx-os.local
servers:
  - url: http://localhost:8081
    description: Local Docker Compose
...
paths:
...
  /api/v1/me:
    get:
      operationId: getCurrentUser
      description: Trả danh tính và realm role đã xác minh từ bearer access token.
      tags: [Identity]
      responses:
        "200":
          description: Principal được suy ra từ access token hợp lệ.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests:
    post:
      operationId: createPurchaseRequest
      description: Tạo phiếu mua sắm DRAFT cho user và phòng ban lấy từ access token/profile nghiệp vụ.
      tags: [Procurement]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseRequest"
      responses:
        "201":
          description: Phiếu mua sắm đã được tạo.
          headers:
            Location:
              description: URI của phiếu vừa tạo.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
    get:
      operationId: listPurchaseRequests
      description: Liệt kê phiếu theo scope của user, có phân trang và lọc trạng thái.
      tags: [Procurement]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          description: Lọc theo trạng thái workflow.
          schema:
            $ref: "#/components/schemas/PurchaseRequestStatus"
      responses:
        "200":
          description: Trang kết quả trong scope được phép.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequestPage"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}:
    get:
      operationId: getPurchaseRequest
      description: Trả chi tiết phiếu và các dòng hàng nếu phiếu thuộc data scope của user.
      tags: [Procurement]
      parameters:
        - name: requestId
          in: path
          required: true
          description: UUID của phiếu mua sắm.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Chi tiết phiếu mua sắm.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
    patch:
      operationId: updatePurchaseRequest
      description: Cập nhật nội dung phiếu DRAFT hoặc CHANGES_REQUESTED do chính user tạo, có optimistic locking.
      tags: [Procurement]
      parameters:
        - name: requestId
          in: path
          required: true
          description: UUID của phiếu mua sắm.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseRequest"
      responses:
        "200":
          description: Phiếu đã được cập nhật và tăng version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}/transitions:
    post:
      operationId: transitionPurchaseRequest
      description: Thực hiện một hành động workflow theo role, data scope và trạng thái hiện tại.
      tags: [Procurement]
      parameters:
        - name: requestId
          in: path
          required: true
          description: UUID của phiếu mua sắm.
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          description: Khóa duy nhất cho một ý định transition; gửi lại cùng khóa sẽ nhận cùng kết quả.
          schema:
            type: string
            minLength: 8
            maxLength: 255
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{7,254}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionPurchaseRequest"
      responses:
        "200":
          description: Transition đã được áp dụng hoặc phát lại an toàn từ cùng Idempotency-Key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}/timeline:
    get:
      operationId: getPurchaseRequestTimeline
      description: Trả lịch sử nghiệp vụ append-only nếu phiếu thuộc data scope của user.
      tags: [Procurement]
      parameters:
        - name: requestId
          in: path
          required: true
          description: UUID của phiếu mua sắm.
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Trang timeline mới nhất trước; không chứa metadata hoặc idempotency key nội bộ.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequestTimelinePage"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}/attachments:
    get:
      operationId: listPurchaseRequestAttachments
      description: Liệt kê metadata tài liệu và quy tắc bắt buộc theo cùng data scope với phiếu.
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/PurchaseRequestId"
      responses:
        "200":
          description: Danh sách tài liệu đang hoạt động và kết quả kiểm tra quy tắc.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequestAttachmentList"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
    post:
      operationId: uploadPurchaseRequestAttachment
      description: Chủ phiếu tải tài liệu lên khi phiếu còn cho phép chỉnh sửa.
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/PurchaseRequestId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [documentType, file]
              properties:
                documentType:
                  $ref: "#/components/schemas/AttachmentDocumentType"
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Tệp và metadata đã được lưu thành công.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequestAttachment"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}/attachments/{attachmentId}:
    delete:
      operationId: deletePurchaseRequestAttachment
      description: Chủ phiếu xóa tài liệu khi phiếu còn DRAFT hoặc CHANGES_REQUESTED.
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/PurchaseRequestId"
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "204":
          description: Tệp đã được xóa khỏi kho và metadata được đánh dấu DELETED.
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}/attachments/{attachmentId}/content:
    get:
      operationId: downloadPurchaseRequestAttachment
      description: Tải nội dung sau khi backend kiểm tra data scope và checksum SHA-256.
      tags: [Documents]
      parameters:
        - $ref: "#/components/parameters/PurchaseRequestId"
        - $ref: "#/components/parameters/AttachmentId"
      responses:
        "200":
          description: Nội dung tệp.
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /api/v1/purchase-requests/{requestId}/budget-check:
    get:
      operationId: checkPurchaseRequestBudget
      description: Returns the active allocation and reservation result for a purchase request in the caller's data scope.
      tags: [Budgets]
      parameters:
        - name: requestId
          in: path
          required: true
          description: Purchase request UUID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Budget availability or reservation state for the purchase request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetCheck"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/budgets/summary:
    get:
      operationId: getBudgetSummary
      description: Returns an active budget allocation. Finance and audit can inspect their organization; managers are limited to their department cost center.
      tags: [Budgets]
      parameters:
        - name: costCenter
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
        - name: currency
          in: query
          required: true
          schema:
            type: string
            pattern: "^[A-Z]{3}$"
      responses:
        "200":
          description: Active allocation totals after reservations and commitments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetSummary"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/budgets/dashboard:
    get:
      operationId: getBudgetDashboard
      description: Returns active allocations, currency totals, reservations and adjustment history for finance or audit in the caller's organization.
      tags: [Budgets]
      responses:
        "200":
          description: Organization-scoped budget management dashboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetDashboard"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/budgets/allocations/{allocationId}:
    patch:
      operationId: adjustBudgetAllocation
      description: Adjusts the total allocation. Finance only; audit remains read-only.
      tags: [Budgets]
      parameters:
        - name: allocationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 8
            maxLength: 255
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{7,254}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBudgetAllocation"
      responses:
        "200":
          description: Updated allocation or an idempotent replay of the adjustment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetAllocation"
        "400":
          $ref: "#/components/responses/Problem"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "415":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/reports/procurement:
    get:
      operationId: getProcurementReport
      description: Trả dashboard vận hành từ curated reporting views. Chỉ finance, auditor và dx_admin được truy cập.
      tags: [Reports]
      parameters:
        - name: from
          in: query
          description: Ngày bắt đầu, mặc định là 29 ngày trước ngày hiện tại.
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Ngày kết thúc, mặc định là ngày hiện tại. Khoảng thời gian tối đa 367 ngày.
          schema:
            type: string
            format: date
        - name: departmentId
          in: query
          description: UUID phòng ban.
          schema:
            type: string
            format: uuid
        - name: costCenter
          in: query
          description: Cost center chuẩn hóa thành chữ hoa.
          schema:
            type: string
            minLength: 2
            maxLength: 100
            pattern: "^[A-Z0-9][A-Z0-9._-]{1,99}$"
        - name: currency
          in: query
          description: Mã tiền tệ ISO dạng ba chữ cái.
          schema:
            type: string
            pattern: "^[A-Z]{3}$"
      responses:
        "200":
          description: KPI, xu hướng, phân bố trạng thái, phòng ban và sử dụng ngân sách theo bộ lọc.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcurementReportDashboard"
        "401":
          $ref: "#/components/responses/Problem"
        "403":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "500":
          $ref: "#/components/responses/Problem"
  /api/v1/me/notifications:
    get:
      operationId: listMyNotifications
      description: Liệt kê notification đã materialize theo principal hiện tại và trạng thái đã đọc.
      tags: [Notifications]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: unreadOnly
          in: query
          schema: { type: boolean, default: false }
      responses:
        "200":
          description: Notification theo đúng user/role/department scope.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPage" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /api/v1/me/notifications/read-all:
    post:
      operationId: markAllMyNotificationsRead
      description: Đánh dấu toàn bộ notification thuộc principal hiện tại là đã đọc.
      tags: [Notifications]
      responses:
        "200":
          description: Số notification vừa đánh dấu đã đọc.
          content:
            application/json:
              schema:
                type: object
                required: [updated]
                properties: { updated: { type: integer, format: int64 } }
        "401": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Problem:
      description: Lỗi theo RFC 9457 Problem Details.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  parameters:
...
    Page:
      name: page
      in: query
      description: Số trang, bắt đầu từ 1.
      schema:
        type: integer
        minimum: 1
        default: 1
...
  schemas:
...
    CurrentUser:
      type: object
      additionalProperties: false
      required: [subject, username, roles]
      properties:
        subject:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            type: string
...
    CreatePurchaseRequest:
      type: object
      additionalProperties: false
      required: [title, reason, currency, costCenter, items]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 255
        reason:
          type: string
          minLength: 10
          maxLength: 5000
        currency:
          type: string
          pattern: "^[A-Z]{3}$"
          examples: [VND]
        costCenter:
          type: string
          minLength: 1
          maxLength: 100
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/CreatePurchaseRequestItem"
...
    AttachmentDocumentType:
      type: string
      enum: [QUOTATION, SPECIFICATION, CONTRACT, OTHER]
...
security:
  - bearerAuth: []